try to separate all the subworld
